remove: SurveyAnswer.Answer 제거(#102)
