+++ src/main/java/com/dnd/namuiwiki/domain/survey/model/entity/Survey.java
@@ -1,13 +1,19 @@
 package com.dnd.namuiwiki.domain.survey.model.entity;
 
+import com.dnd.namuiwiki.common.exception.ApplicationErrorException;
+import com.dnd.namuiwiki.common.exception.ApplicationErrorType;
 import com.dnd.namuiwiki.common.model.BaseTimeEntity;
+import com.dnd.namuiwiki.domain.question.dto.QuestionDto;
 import com.dnd.namuiwiki.domain.question.entity.Question;
+import com.dnd.namuiwiki.domain.question.type.QuestionType;
 import com.dnd.namuiwiki.domain.survey.type.AnswerType;
 import com.dnd.namuiwiki.domain.survey.type.Period;
 import com.dnd.namuiwiki.domain.survey.type.Relation;
 import com.dnd.namuiwiki.domain.user.entity.User;
+import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
+import lombok.NoArgsConstructor;
 import org.springframework.data.annotation.Id;
 import org.springframework.data.mongodb.core.mapping.Document;
 import org.springframework.data.mongodb.core.mapping.DocumentReference;
@@ -40,6 +46,8 @@ public class Survey extends BaseTimeEntity {
 
     @Getter
     @Builder
+    @NoArgsConstructor
+    @AllArgsConstructor
     public static class Answer {
 
         @DocumentReference(collection = "questions", lazy = true)
@@ -47,6 +55,60 @@ public class Survey extends BaseTimeEntity {
         private AnswerType type;
         private Object answer;
         private String reason;
+
+        public Answer(QuestionDto question, AnswerType type, Object answer, String reason) {
+            validateAnswerType(question.getType(), type);
+            validateReasonRequired(question.isReasonRequired(), reason);
+            boolean shouldBeNumericAnswer = question.getType().isNumericType() && type.isManual();
+            if (shouldBeNumericAnswer) {
+                validateAnswerObjectType(answer, Integer.class);
+                long longAnswer = Long.parseLong(answer.toString());
+                if (question.getName().isBorrowingLimit()) {
+                    validateBorrowingLimit(longAnswer);
+                }
+                this.answer = longAnswer;
+            } else {
+                validateAnswerObjectType(answer, String.class);
+                this.answer = answer;
+            }
+
+            this.question = question.toEntity();
+            this.type = type;
+            this.reason = reason;
+        }
+
+        private void validateBorrowingLimit(long longAnswer) {
+            if (!(0 <= longAnswer && longAnswer <= 1_000_000_000)) {
+                throw new ApplicationErrorException(ApplicationErrorType.INVALID_BORROWING_LIMIT);
+            }
+        }
+
+        public Survey.Answer toEntity() {
+            return Survey.Answer.builder()
+                    .question(question)
+                    .type(type)
+                    .answer(answer)
+                    .reason(reason)
+                    .build();
+        }
+
+        private void validateAnswerObjectType(Object answer, Class<?> clazz) {
+            if (!clazz.isInstance(answer)) {
+                throw new ApplicationErrorException(ApplicationErrorType.NOT_INTEGER_ANSWER);
+            }
+        }
+
+        private void validateReasonRequired(boolean reasonRequired, String reason) {
+            if (reasonRequired && reason == null) {
+                throw new ApplicationErrorException(ApplicationErrorType.ANSWER_REASON_REQUIRED);
+            }
+        }
+
+        private void validateAnswerType(QuestionType questionType, AnswerType answerType) {
+            if (!questionType.containsAnswerType(answerType)) {
+                throw new ApplicationErrorException(ApplicationErrorType.NOT_ALLOWED_ANSWER_TYPE);
+            }
+        }
 
     }
